requirements.txt
keras==2.13.1
numpy~=1.24.3
pandas==2.0.0
tensorflow==2.13.0
scikit-learn~=1.3.0
google-cloud-pubsub==2.18.1
Pillow==10.0.0